requirements.txt
Django==3.2.3
djangorestframework
django-allauth
dj-rest-auth
postgres
django-bootstrap3
requests
djangorestframework-simplejwt
coverage
